Organize host integration test ownership
- Separate registration, signature, and execution scenarios by role.
- Keep public host examples local and improve explanation readability.
